Revert "Send logs to stdout/stderr"
This reverts commit 6d94aa7f861a7d820aa47f14de6190d97efd5e71, because it wasn't
really needed - Kong environment variables such as `KONG_PROXY_ACCESS_LOG` are
enough for logging to stdout/stderr when needed. The problem in our case was
that the NGINX process that did the logging wasn't started as the main
container process, but was started by `kong start` and so it had its own
stdout, which wasn't collected by Docker.